requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.11
requests==2.31.0
pytest==7.4.0
pymongo==4.5.0
python-dotenv==1.0.0
groq==0.4.2